[reduce-deps] move toslice to fun, separate seq routines, add tests for toslice
