+++ assert_test.go
@@ -16,15 +16,6 @@ type User struct {
 	pass  Pass
 }
 
-func toSlice[T any](seq fun.Seq[T]) []T {
-	var xs []T
-	seq(func(x T) bool {
-		xs = append(xs, x)
-		return true
-	})
-	return xs
-}
-
 func TestDiffImpl(t *testing.T) {
 	// must not panic on comparing structs in private field User.pass
 	expected := []diffLine{
@@ -32,7 +23,7 @@ func TestDiffImpl(t *testing.T) {
 		{expected: "b", actual: "e", selector: ".pass.Payload"},
 		{expected: "c", actual: "f", selector: ".pass.salt"},
 	}
-	actual := toSlice(diffImpl("",
+	actual := fun.ToSlice(diffImpl("",
 		User{"a", Pass{"b", "c"}},
 		User{"d", Pass{"e", "f"}},
 	))
